move SimEngine -> SimEngineBase, so that netlist-specific logging can be created in a child class. Move core infra around agents (AgentWallet, AgentBase, AgentDict) to engine/.
