+++ assets/agents/RouterAgent.py
@@ -5,11 +5,11 @@ from enforce_typing import enforce_types
 from typing import List
 import math
 
-from assets.agents.BaseAgent import BaseAgent
+from engine.AgentBase import AgentBase
 from util.constants import S_PER_MONTH
     
 @enforce_types
-class RouterAgent(BaseAgent):
+class RouterAgent(AgentBase):
     def __init__(self, name: str, USD: float, OCEAN: float,
                  receiving_agents : dict):
         """receiving_agents -- [agent_n_name] : method_for_%_going_to_agent_n
