+++ assets/agents/StakerspeculatorAgent.py
@@ -1,13 +1,13 @@
 from enforce_typing import enforce_types
 import random
 
-from assets.agents.BaseAgent import BaseAgent
+from engine.AgentBase import AgentBase
 from web3engine import bfactory, bpool, btoken, datatoken, dtfactory
 from web3tools.web3util import toBase18
 from util import constants
                     
 @enforce_types
-class StakerspeculatorAgent(BaseAgent):
+class StakerspeculatorAgent(AgentBase):
     """Speculates by staking and unstaking"""
     
     def __init__(self, name: str, USD: float, OCEAN: float):
