+++ assets/agents/test/test_RouterAgent.py
@@ -1,6 +1,6 @@
 from enforce_typing import enforce_types
 
-from assets.agents.BaseAgent import BaseAgent
+from engine.AgentBase import AgentBase
 from assets.agents.RouterAgent import RouterAgent
 from engine import SimState, SimStrategy
 from util.constants import S_PER_DAY, S_PER_MONTH
@@ -12,7 +12,7 @@ def test1():
     ss.time_step = S_PER_DAY
     state = SimState.SimState(ss)
 
-    class SimpleAgent(BaseAgent):
+    class SimpleAgent(AgentBase):
         def takeStep(self, state):
             pass
     state.agents["a1"] = a1 = SimpleAgent("a1", 0.0, 0.0)
