+++ assets/agents/test/test_StakerspeculatorAgent.py
@@ -1,7 +1,7 @@
 from enforce_typing import enforce_types
 import pytest
 
-from assets.agents.AgentDict import AgentDict
+from engine.AgentDict import AgentDict
 from assets.agents.PoolAgent import PoolAgent
 from assets.agents.StakerspeculatorAgent import StakerspeculatorAgent
 from util.constants import S_PER_HOUR
