+++ engine/SimEngineBase.py
@@ -1,0 +1,119 @@
+import logging
+log = logging.getLogger('master')
+
+from enforce_typing import enforce_types
+import os
+
+from util import valuation
+from util.constants import S_PER_MIN, S_PER_HOUR, S_PER_DAY, S_PER_MONTH, S_PER_YEAR
+
+@enforce_types
+class SimEngineBase(object):
+    """
+    @description
+      Runs a simulation.
+      
+    @attributes
+      state - child of SimState
+      output_dir -- directory of where results are stored
+    """
+
+    def __init__(self, state, output_dir: str):
+        self.state = state
+        self.output_dir = output_dir
+        self.output_csv = "data.csv" #magic number
+        
+    def run(self):
+        """
+        @description
+          Runs the simulation!  This is the main work routine.
+        
+        @return
+           <<none>> but it continually generates an output csv output_dir
+        """
+        log.info("Begin.")
+        log.info(str(self.state.ss) + "\n")
+
+        while True:
+            self.takeStep()
+            if self.doStop():
+                break
+            self.state.tick += 1 #could be e.g. 10 or 100 or ..
+        log.info("Done")
+
+    def takeStep(self) -> None:
+        """Run one tick, updates self.state"""
+        log.debug("=============================================")
+        log.debug("Tick=%d: begin" % (self.state.tick))
+        
+        if (self.elapsedSeconds() % S_PER_DAY) == 0:
+            str_data, csv_data = self.createLogData()
+            log.info(str_data)
+            self.logToCsv(csv_data)
+                
+        #main work
+        self.state.takeStep()
+        
+        log.debug("=============================================")
+        log.debug("Tick=%d: done" % self.state.tick)
+
+    def createLogData(self):
+        #To log more, create a child class that re-implements this method
+        
+        state = self.state
+        ss = state.ss
+        kpis = state.kpis
+
+        s = [] #for console logging
+        dataheader = [] # for csv logging: list of string
+        datarow = [] #for csv logging: list of float
+
+        s += ["Tick=%d" % (state.tick)]
+        dataheader += ["Tick"]
+        datarow += [state.tick]
+
+        es = float(self.elapsedSeconds())
+        emi, eh, ed, emo, ey = es/S_PER_MIN, es/S_PER_HOUR, es/S_PER_DAY, \
+                               es/S_PER_MONTH,es/S_PER_YEAR
+        s += [" (%.1f h, %.1f d, %.1f mo)" % \
+              (eh, ed, emo)] 
+        dataheader += ["Second", "Min", "Hour", "Day", "Month", "Year"]
+        datarow += [es, emi, eh, ed, emo, ey]
+
+        # (here's where you'd log more things)
+        
+        #package results up
+        str_data = "".join(s)
+        csv_data = (dataheader, datarow)
+        return str_data, csv_data
+
+    def logToCsv(self, csv_data) -> None:
+        if self.output_dir is None:
+            return
+                
+        if not os.path.exists(self.output_dir):
+            os.mkdir(self.output_dir)
+            
+        full_filename = os.path.join(self.output_dir, self.output_csv)
+        (dataheader, datarow) = csv_data
+        
+        #if needed, create file and add header
+        if not os.path.exists(full_filename):
+            with open(full_filename,'w+') as f:
+                f.write(", ".join(dataheader) + "\n")
+            
+        #add in row
+        datarow_s = ['%g' % dataval for dataval in datarow]
+        with open(full_filename,'a+') as f:
+            f.write(", ".join(datarow_s) + "\n")
+
+    def elapsedSeconds(self) -> int:
+        return self.state.tick * self.state.ss.time_step
+
+    def doStop(self) -> bool:
+        if self.state.tick >= self.state.ss.max_ticks:
+            log.info("Stop: tick (%d) >= max" % self.state.tick)
+            return True
+        
+        return False
+
