+++ engine/SimStateBase.py
@@ -1,5 +1,5 @@
 from enforce_typing import enforce_types
-from assets.agents.AgentDict import AgentDict
+from engine.AgentDict import AgentDict
 
 @enforce_types
 class SimStateBase(object):
